hparams: swap demos, renaming original to “minimal”
Summary:
Now, the default demo is the new, ML-based one. The original demo is
preserved because it’s orders of magnitudes faster.

Test Plan:
`bazel run`ning each demo still works.

wchargin-branch: hparams-ml-demo
